Created shared base class for Disguise Modifiers
